+++ game/settings/settings.py
@@ -324,6 +324,19 @@ class Settings:
             "modifications."
         ),
     )
+    reset_simulation_on_abort: bool = boolean_option(
+        "Reset mission to pre-take off conditions on abort (experimental)",
+        page=MISSION_GENERATOR_PAGE,
+        section=GAMEPLAY_SECTION,
+        default=False,
+        detail=(
+            "If enabled, the fast-forward effects will be rewound when aborting take "
+            "off. <strong>DO NOT USE THIS WITH AUTO-RESOLVE ENABLED.</strong> Lost "
+            "aircraft will not be recovered. This option is experimental and may not "
+            "work. It is always safer to reload your save after abort when using fast-"
+            "forward."
+        ),
+    )
     player_mission_interrupts_sim_at: Optional[StartType] = choices_option(
         "Player missions interrupt fast forward",
         page=MISSION_GENERATOR_PAGE,